move extension types and corresponding document methods to their own files
