+++ pkg/ast/ast_input_object_type_definition.go
@@ -72,14 +72,6 @@ func (d *Document) InputObjectTypeDefinitionInputValueDefinitionByName(definitio
 	return -1
 }
 
-func (d *Document) InputObjectTypeExtensionHasDirectives(ref int) bool {
-	return d.InputObjectTypeExtensions[ref].HasDirectives
-}
-
-func (d *Document) InputObjectTypeExtensionHasInputFieldsDefinition(ref int) bool {
-	return d.InputObjectTypeDefinitions[ref].HasInputFieldsDefinition
-}
-
 func (d *Document) InputObjectTypeDefinitionNameBytes(ref int) ByteSlice {
 	return d.Input.ByteSlice(d.InputObjectTypeDefinitions[ref].Name)
 }
@@ -97,31 +89,4 @@ func (d *Document) InputObjectTypeDefinitionDescriptionBytes(ref int) ByteSlice 
 
 func (d *Document) InputObjectTypeDefinitionDescriptionString(ref int) string {
 	return unsafebytes.BytesToString(d.InputObjectTypeDefinitionNameBytes(ref))
-}
-
-type InputObjectTypeExtension struct {
-	ExtendLiteral position.Position
-	InputObjectTypeDefinition
-}
-
-func (d *Document) InputObjectTypeExtensionNameBytes(ref int) ByteSlice {
-	return d.Input.ByteSlice(d.InputObjectTypeExtensions[ref].Name)
-}
-
-func (d *Document) InputObjectTypeExtensionNameString(ref int) string {
-	return unsafebytes.BytesToString(d.Input.ByteSlice(d.InputObjectTypeExtensions[ref].Name))
-}
-
-func (d *Document) ExtendInputObjectTypeDefinitionByInputObjectTypeExtension(inputObjectTypeDefinitionRef, inputObjectTypeExtensionRef int) {
-	if d.InputObjectTypeExtensionHasDirectives(inputObjectTypeExtensionRef) {
-		d.InputObjectTypeDefinitions[inputObjectTypeDefinitionRef].Directives.Refs = append(d.InputObjectTypeDefinitions[inputObjectTypeDefinitionRef].Directives.Refs, d.InputObjectTypeExtensions[inputObjectTypeExtensionRef].Directives.Refs...)
-		d.InputObjectTypeDefinitions[inputObjectTypeDefinitionRef].HasDirectives = true
-	}
-
-	if d.InputObjectTypeExtensionHasInputFieldsDefinition(inputObjectTypeExtensionRef) {
-		d.InputObjectTypeDefinitions[inputObjectTypeDefinitionRef].InputFieldsDefinition.Refs = append(d.InputObjectTypeDefinitions[inputObjectTypeDefinitionRef].InputFieldsDefinition.Refs, d.InputObjectTypeExtensions[inputObjectTypeExtensionRef].InputFieldsDefinition.Refs...)
-		d.InputObjectTypeDefinitions[inputObjectTypeDefinitionRef].HasInputFieldsDefinition = true
-	}
-
-	d.Index.MergedTypeExtensions = append(d.Index.MergedTypeExtensions, Node{Ref: inputObjectTypeExtensionRef, Kind: NodeKindInputObjectTypeExtension})
 }
