+++ pkg/ast/ast_interface_type_definition.go
@@ -20,14 +20,6 @@ type InterfaceTypeDefinition struct {
 	FieldsDefinition    FieldDefinitionList // optional, e.g. { name: String }
 }
 
-func (d *Document) InterfaceTypeExtensionHasDirectives(ref int) bool {
-	return d.InterfaceTypeExtensions[ref].HasDirectives
-}
-
-func (d *Document) InterfaceTypeExtensionHasFieldDefinitions(ref int) bool {
-	return d.InterfaceTypeExtensions[ref].HasFieldDefinitions
-}
-
 func (d *Document) InterfaceTypeDefinitionNameBytes(ref int) ByteSlice {
 	return d.Input.ByteSlice(d.InterfaceTypeDefinitions[ref].Name)
 }
@@ -45,31 +37,4 @@ func (d *Document) InterfaceTypeDefinitionDescriptionBytes(ref int) ByteSlice {
 
 func (d *Document) InterfaceTypeDefinitionDescriptionString(ref int) string {
 	return unsafebytes.BytesToString(d.InterfaceTypeDefinitionDescriptionBytes(ref))
-}
-
-type InterfaceTypeExtension struct {
-	ExtendLiteral position.Position
-	InterfaceTypeDefinition
-}
-
-func (d *Document) InterfaceTypeExtensionNameBytes(ref int) ByteSlice {
-	return d.Input.ByteSlice(d.InterfaceTypeExtensions[ref].Name)
-}
-
-func (d *Document) InterfaceTypeExtensionNameString(ref int) string {
-	return unsafebytes.BytesToString(d.Input.ByteSlice(d.InterfaceTypeExtensions[ref].Name))
-}
-
-func (d *Document) ExtendInterfaceTypeDefinitionByInterfaceTypeExtension(interfaceTypeDefinitionRef, interfaceTypeExtensionRef int) {
-	if d.InterfaceTypeExtensionHasFieldDefinitions(interfaceTypeExtensionRef) {
-		d.InterfaceTypeDefinitions[interfaceTypeDefinitionRef].FieldsDefinition.Refs = append(d.InterfaceTypeDefinitions[interfaceTypeDefinitionRef].FieldsDefinition.Refs, d.InterfaceTypeExtensions[interfaceTypeExtensionRef].FieldsDefinition.Refs...)
-		d.InterfaceTypeDefinitions[interfaceTypeDefinitionRef].HasFieldDefinitions = true
-	}
-
-	if d.InterfaceTypeExtensionHasDirectives(interfaceTypeExtensionRef) {
-		d.InterfaceTypeDefinitions[interfaceTypeDefinitionRef].Directives.Refs = append(d.InterfaceTypeDefinitions[interfaceTypeDefinitionRef].Directives.Refs, d.InterfaceTypeExtensions[interfaceTypeExtensionRef].Directives.Refs...)
-		d.InterfaceTypeDefinitions[interfaceTypeDefinitionRef].HasDirectives = true
-	}
-
-	d.Index.MergedTypeExtensions = append(d.Index.MergedTypeExtensions, Node{Ref: interfaceTypeExtensionRef, Kind: NodeKindInterfaceTypeExtension})
 }
