+++ pkg/ast/ast_scalar_type_definition.go
@@ -27,29 +27,3 @@ func (d *Document) ScalarTypeDefinitionNameBytes(ref int) ByteSlice {
 func (d *Document) ScalarTypeDefinitionNameString(ref int) string {
 	return unsafebytes.BytesToString(d.Input.ByteSlice(d.ScalarTypeDefinitions[ref].Name))
 }
-
-type ScalarTypeExtension struct {
-	ExtendLiteral position.Position
-	ScalarTypeDefinition
-}
-
-func (d *Document) ScalarTypeExtensionHasDirectives(ref int) bool {
-	return d.ScalarTypeExtensions[ref].HasDirectives
-}
-
-func (d *Document) ScalarTypeExtensionNameBytes(ref int) ByteSlice {
-	return d.Input.ByteSlice(d.ScalarTypeExtensions[ref].Name)
-}
-
-func (d *Document) ScalarTypeExtensionNameString(ref int) string {
-	return unsafebytes.BytesToString(d.Input.ByteSlice(d.ScalarTypeExtensions[ref].Name))
-}
-
-func (d *Document) ExtendScalarTypeDefinitionByScalarTypeExtension(scalarTypeDefinitionRef, scalarTypeExtensionRef int) {
-	if d.ScalarTypeExtensionHasDirectives(scalarTypeExtensionRef) {
-		d.ScalarTypeDefinitions[scalarTypeDefinitionRef].Directives.Refs = append(d.ScalarTypeDefinitions[scalarTypeDefinitionRef].Directives.Refs, d.ScalarTypeExtensions[scalarTypeExtensionRef].Directives.Refs...)
-		d.ScalarTypeDefinitions[scalarTypeDefinitionRef].HasDirectives = true
-	}
-
-	d.Index.MergedTypeExtensions = append(d.Index.MergedTypeExtensions, Node{Ref: scalarTypeExtensionRef, Kind: NodeKindScalarTypeExtension})
-}
