+++ pkg/ast/ast_schema_extension.go
@@ -1,0 +1,8 @@
+package ast
+
+import "github.com/jensneuse/graphql-go-tools/pkg/lexer/position"
+
+type SchemaExtension struct {
+	ExtendLiteral position.Position
+	SchemaDefinition
+}
