+++ pkg/ast/index.go
@@ -4,20 +4,20 @@ package ast
 type Index struct {
 	// QueryTypeName is the name of the query type on the schema Node
 	// schema { query: Query }
-	QueryTypeName           ByteSlice
+	QueryTypeName ByteSlice
 	// MutationTypeName is the name of the mutation type on the schema Node
 	// schema { mutation: Mutation }
-	MutationTypeName        ByteSlice
+	MutationTypeName ByteSlice
 	// SubscriptionTypeName is the name of the subscription type on the schema Node
 	// schema { subscription: Subscription }
-	SubscriptionTypeName    ByteSlice
+	SubscriptionTypeName ByteSlice
 	// Nodes is a list of all root nodes in a schema definition
 	// The map key is the result of the xxhash algorithm from the Node name.
-	Nodes                   map[uint64]Node
+	Nodes map[uint64]Node
 	// ReplacedFragmentSpreads is a list of references (slice indices) of all FragmentSpreads that got replaced during normalization.
 	ReplacedFragmentSpreads []int
 	// MergedTypeExtensions is a list of Nodes (Node kind + reference) that got merged during type extension merging.
-	MergedTypeExtensions    []Node
+	MergedTypeExtensions []Node
 }
 
 // Reset empties the Index
